pyproject.toml
[project]
name = "semantic-search-amazon"
version = "0.1.0"
description = ""
authors = [
    {name = "Aneesh Naresh",email = "an524@cornell.edu"}
]
readme = "README.md"
requires-python = ">=3.11, <=3.12.10"
dependencies = [
    "jupyterlab (>=4.4.4,<5.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "superlinked (>=29.1.0,<30.0.0)",
    "google-genai (>=1.24.0,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
